recorder/pyproject.toml
[project]
name = "recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "kafka-python>=2.0.2",
]
